transDjango/requirements.txt
Django==1.10.5
django-leaflet==0.19.0
djangorestframework==3.5.4
djangorestframework-gis==0.11
psycopg2==2.6.2
requests==2.13.0
six==1.10.0
